Add ETag support for customizability
Updated readme regarding new functionality and new properties possible
to set in the setup of shapeshifter-middleware.

It's possible to set headers if a call to an endpoint already has an
ETag stored in the middleware, then we can take proper action and
respond to it however we want.

Otherwise we have a fallback that defaults to:

```
{
    'If-None-Match': <matchedETag_here>,
    'Cache-Control': 'private, must-revalidate'
}
```

We also emit an action `dispatchETagCreationType` with the value set, if
any is set. And only when a new call is made and the server responses
with an ETag we can store.

Also introduced a new property to Action(s), `ETagCallback`. It expects
either an object to be emitted or a function that receives the matching
ETag along with normal store operations. The function is expected to
return an object that will hold the headers we pass to the ajax call
anytime we have a matching ETag stored for the specific call to the
endpoint made.