Add middleware opt to emit neutral type
A situation where you might want the middleware to emit the neutral type could
be e.g. once you start loading something big you want the application to set its
state to `loading: true` until it's finished loading.

This is what the neutral type is for, to react once a request is made.