Switch curried args placement
In order to make the curried functions more reusable I had to switch placements
of when the arguments were passed in. This to make some sort of boilerplate for
later when I will need to call the very same method again and still having the
`store` and `next` arguments already prepended.