Correctly set body for post, put, patch & delete
Before I was only correctly setting the request body for `post`, for
some reason. Letting all other request methods make use of the config
and the `params`-property which sets the URL parameters.. completely off
so this could be a hiccup when someone upgrades the lib to the latest
version.