Avoid passing undefined actions
I was calling next() with the reason thinking that I would want to pass
`undefined` to reach the store and therefore trigger the error message
saying can't be `undefined`. However this caused problems when other
middlewares doesn't care for this and therefore crashes.

However we will still log an error about it unless it's in NODE_ENV=test