Consider other actions of not type Object
I was only passing along actions to `next()` if the constructor of the
action was of type `Object`. This is wrong and meant that e.g.
`redux-thunk` actions which is of type `Function` wouldn't be passed to
`next()` but instead tried to be used as if it was a valid shapeshifter
action.