Add opt to get full response object in success()
Previously I was only interested in returning the `response.data` after
a successful request, however there are cases where one wants to
actually look at e.g. the headers of the response.
The workaround to do so was to `.then()`-chain the action and from there
retrieve it, however this isn't well documented (if at all) and might
cause problems for whoever uses this middleware who can't spend time
reading through the source code.