Create log module and implement it
I was making use of `console` object and its methods throughout the
project and most of the time I had to guard against them being called in
tests. Now instead we make use of a utility logger function that would
only log through the `console` object if the `NODE_ENV !== 'test'`. This
also allows me to instead mock the log module and check if it was
called as the log module has its own tests that ensure it pipes through
to the `console` object so I wouldn't have to worry about that