Fix return value of chained middlewares
By simply calling `next(action)` it will continue the chain of
middlewares after shapeshifter. However the return value is of high
importance here as well, I was returning `undefined` implicitly and this
would cause a problem to rise in the following example:

```
/**
 * Middlewares defined as such
 *
 * applyMiddleware(
 *  shapeshifterMiddleware,
 *  thunk,
 * )
 */

dispatch(thunkActionAsPromise)
    .then(() => {})
```

where the result of `dispatch(thunkActionAsPromise)` would return
`undefined` to the next middleware in the chain, hence crashign when
trying to access `.then`. This because shapeshifter was defined before
redux-thunk and its return value would be `undefined`. However the call
to the 2nd middleware in the chain would still be fired but the
reference to it would break.