Swallow 304 errors
If axios sees a 304 status code response from the server it's going to
interpret that as an Error. This is vague error-wise and instead I
transform it into a `ResponseNotModified`-error and keep the stack trace
from the old one.
Once we know the error is `ResponseNotModified` we can act accordingly
and try calling the ETagCallback function, if provided, or dispatch the
object provided. At the end I will bail early without dispatching
`failure()` and logging either warn or error.