Refactor handleResponse function
`handleResponse.js` is refactored and also starts adhering to new code
style, which is why the diffs are a bit bigger in this commit and
unclear than usual.

As a new convention I started with this also now takes a context with
what it needs from the middleware scope and throws on invalid
`response.data`.

Tests were written now for this.