Decouple repeater functionality into own module
To increase testability of the repeater functionality and reduce noise
in the `middleware.js` file this piece of code is moved into its own
module. The `axios-mock-adapter` was also introduced to easier work with
axios requests during testing.

Also added some additional "fixes" to the `test/index.js` in order to
show in the git history what was changed and that the old tests still
passes as the library has been dependent on those tests. However these
tests are now superfluous and will be removed in due time.