Resolve vulnerablities in packages
Upgraded `npm` to its latest version to resolve some of the
vulnerabilities found. However this didn't cover all of them and even
though it's the latest it still had some dependencies with
vulnerabilities. Because of this I used `yarn`'s `resolution` property
in the `package.json` to correct the version of the dependencies that
still had flaws in them as well as other packages in order to resolve
the audit warnings.