Apply linter on entire project
After adding the linter I've run it on the entire project using the
`--fix` flag and also done manual fixes that eslint couldn't handle
itself.

Thanks to the linter it appeared that there were few dependency cycles
within the project itself. Reason was that I was importing
`src/middleware.js` in other modules in order to get the middleware
options. To resolve this I created `src/options.js` that exports the
default middleware options.
This was however not the end as I still had two tests not passing. They
were expecting a mutated state of the middleware options, user options
mixed with defaults, but there was no way of accessing this mutated
state. That's why I introduced a so-so fix for now where I define the
returned function (which is the first function of the curried middleware
functions) with a name (function shapeshifter() { ... }) that I define a
property on with a getter method that returns the internal mutated
state. The tests are now accessing it through that property and are now
passing.

I've also sprinkled `eslint-disable-next-line` here and there where I
actually want to spend time on refactoring and doing it right instead of
doing an ugly fix to have the linter passing.