Rename `prepublish` to `prepare`
Since `npm@5` the `prepublish` script got deprecated so `prepare` and
`prepublishOnly` are instead introduced.

The reason for this is that `prepublish` would run on both `npm publish`
and `npm install`, which is not a wanted behavior for everyone. For me
it works just great though. Because I intend to keep the very same
behavior I simply rename my script-command to its new equivalent.